requirements.txt
langchain==0.3.16
chromadb==0.6.3
langchain-google-genai==2.0.9
google-generativeai==0.8.4
pypdf==5.2.0
streamlit==1.42.0
aiohttp
langchain-core
langchain-text-splitters
langsmith
numpy
pydantic
PyYAML
requests
SQLAlchemy
tenacity
bcrypt
build
chroma-hnswlib
fastapi
grpcio
httpx
importlib-resources
kubernetes
mmh3
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-sdk
orjson
overrides
posthog
pypika
rich
tokenizers
tqdm
typer
typing_extensions
uvicorn
filetype
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
protobuf
altair
blinker
cachetools
click
gitpython
packaging
pandas
pillow
pyarrow
pydeck
watchdog
